pull command creation functions out and implement macos exists and which
