+++ src/bridge/mod.rs
@@ -1,4 +1,5 @@
 mod clipboard;
+mod command;
 pub mod create;
 mod events;
 mod handler;
@@ -6,17 +7,17 @@ mod setup;
 mod tx_wrapper;
 mod ui_commands;
 
-use std::{path::Path, process::Stdio, sync::Arc, thread};
+use std::{process::exit, sync::Arc, thread};
 
-use log::{error, info, warn};
+use log::{error, info};
 use nvim_rs::UiAttachOptions;
-use tokio::process::Command;
 
 use crate::{
     cmd_line::CmdLineSettings, error_handling::ResultPanicExplanation, running_tracker::*,
     settings::*,
 };
 
+use command::create_nvim_command;
 pub use events::*;
 use handler::NeovimHandler;
 use setup::setup_neovide_specific_state;
@@ -56,7 +57,7 @@ async fn start_neovim_runtime() {
         Ok("1") => {} // This is just a guard
         _ => {
             error!("Neovide requires nvim version 0.4 or higher. Download the latest version here https://github.com/neovim/neovim/wiki/Installing-Neovim");
-            std::process::exit(0);
+            exit(0);
         }
     }
 
@@ -97,107 +98,4 @@ async fn start_neovim_runtime() {
         Ok(Ok(())) => {}
     };
     RUNNING_TRACKER.quit("neovim processed failed");
-}
-
-#[cfg(windows)]
-fn set_windows_creation_flags(cmd: &mut Command) {
-    cmd.creation_flags(0x0800_0000); // CREATE_NO_WINDOW
-}
-
-fn build_nvim_cmd_with_args(bin: &str) -> Command {
-    let mut args = vec!["--embed".to_string()];
-    args.extend(SETTINGS.get::<CmdLineSettings>().neovim_args);
-
-    #[cfg(windows)]
-    if SETTINGS.get::<CmdLineSettings>().wsl {
-        let mut cmd = Command::new("wsl");
-        let argstring = format!("{} {}", bin.trim(), args.join(" "));
-        cmd.args(&["$SHELL", "-lc", &argstring]);
-        return cmd;
-    }
-
-    #[cfg(macos)]
-    {
-        let shell = env::var("SHELL").unwrap();
-        let mut cmd = Command::new(shell);
-        let argstring = format!("{} {}", bin.trim(), args.join(" "));
-        cmd.args(&["-lc", &argstring]);
-        return cmd;
-    }
-
-    let mut cmd = Command::new(bin);
-    cmd.args(args);
-    cmd
-}
-
-fn platform_exists(bin: &str) -> bool {
-    #[cfg(windows)]
-    if SETTINGS.get::<CmdLineSettings>().wsl {
-        if let Ok(output) = std::process::Command::new("wsl")
-            .args(&["$SHELL", "-lic"])
-            .arg(format!("exists -x {}", bin))
-            .output()
-        {
-            return output.status.success();
-        } else {
-            error!("wsl exists failed");
-            std::process::exit(1);
-        }
-    }
-    Path::new(&bin).exists()
-}
-
-fn platform_which(bin: &str) -> Option<String> {
-    #[cfg(windows)]
-    if SETTINGS.get::<CmdLineSettings>().wsl {
-        if let Ok(output) = std::process::Command::new("wsl")
-            .args(&["$SHELL", "-lic"])
-            .arg(format!("which {}", bin))
-            .output()
-        {
-            if output.status.success() {
-                return Some(String::from_utf8(output.stdout).unwrap());
-            } else {
-                return None;
-            }
-        }
-    }
-    if let Ok(path) = which::which(bin) {
-        path.into_os_string().into_string().ok()
-    } else {
-        None
-    }
-}
-
-fn build_nvim_cmd() -> Command {
-    if let Some(path) = SETTINGS.get::<CmdLineSettings>().neovim_bin {
-        if platform_exists(&path) {
-            return build_nvim_cmd_with_args(&path);
-        } else {
-            warn!("NEOVIM_BIN is invalid falling back to first bin in PATH");
-        }
-    }
-    if let Some(path) = platform_which("nvim") {
-        build_nvim_cmd_with_args(&path)
-    } else {
-        error!("nvim not found!");
-        std::process::exit(1);
-    }
-}
-
-pub fn create_nvim_command() -> Command {
-    let mut cmd = build_nvim_cmd();
-
-    info!("Starting neovim with: {:?}", cmd);
-
-    #[cfg(not(debug_assertions))]
-    cmd.stderr(Stdio::piped());
-
-    #[cfg(debug_assertions)]
-    cmd.stderr(Stdio::inherit());
-
-    #[cfg(windows)]
-    set_windows_creation_flags(&mut cmd);
-
-    cmd
 }
